initial
Q1: Write a C program that accepts a file name and a substring as command line argument
and prints the number of occurrences of substring in the given file on the console. If the
file does not exist, print some error on the screen.
--------------------------------------------
Q3 : Write a program that reverses all the characters of words that are not vowels in a given
file.